Imported Build and Create Java Project
Imported all the previous built website. This gives us a frame of
refrences as well as acess to all of the images and whatnot.
Created the Java Project that all of the Java code will be writted in.
Began working on the abstract Page class, though did not get very far.