Clean up JavaScript
Removed Pinterest Support.
Created a method in Page.java (and an overwrite in Product.java) calle
getPageIndex() that gets the index of the product. It returns -1 for all
pages other than products. This speeds up the product page JS lookups.
Created a method in Page.java (and an override in Search.java) called
getOnLoadCall() that return the JS function to be called when the page
loads. This allows the function to be different for Search.html.
Cleaned up Javascript.js:
-- Renamed updateHeader(name, add) to updateProductPage(add)
--Renamed addExtraPhotos(name, index) to updateProductPage(name, index).
Also deleted the original updateProductPage(name).
--No new support for searthFunction(add, isBig), emailList(isEmailList)
or any of Search.js. This should happen in future updates.